Move code from RobotContainer to Robot
